move BlogPageContext to utils data
