+++ src/templates/Blog.tsx
@@ -7,15 +7,12 @@ import Layout from "../components/Layout";
 import Pagination from "../components/Pagination";
 import PostCard from "../components/PostCard";
 import SEO from "../components/SEO";
-import { getBlogMetaData, getPostData, postKeyMap } from "../utils/data";
-
-export interface BlogPageContext {
-  type: "Index" | "Category" | "Tag" | "Language";
-  filterValue: string;
-  basePath: string;
-  page: number;
-  numPages: number;
-}
+import {
+  BlogPageContext,
+  getBlogMetaData,
+  getPostData,
+  postKeyMap
+} from "../utils/data";
 
 const Blog: React.FC<
   PageProps<GatsbyTypes.BlogTemplateQuery, BlogPageContext>
